requirements.txt
paho-mqtt==1.6.1
matplotlib==3.8.2